Add arrows
