--- src/styles/Home.css
@@ -1,27 +1,0 @@
-.home-container {
-  margin-left: 10%;
-}
-
-.home-header {
-  min-height: 10vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 5%;
-  margin-bottom: 5%;
-}
-
-.title {
-  font-size: 5rem;
-}
-
-.header-text {
-  color: gray;
-  font-size: 2em;
-}
-
-.header-subtext {
-  color: gray;
-  font-size: 1.5em;
-}
