+++ src/styles/Home.scss
@@ -1,0 +1,63 @@
+$green-color: #97BF11;
+$gray-color: #474747;
+
+.home-container {
+  margin-left: 250px;
+  position: relative;
+
+  .arrows {
+    width: 100%;
+
+    .arrow-left,
+    .arrow-right {
+      position: absolute;
+      width: 100px;
+      background-color: $green-color;
+      font-size: 2.5rem;
+      color: white;
+      cursor: pointer;
+      padding: 20px;
+      transition: all 0.3s ease-in-out;
+      top: 50%;
+      z-index: 5;
+    }
+
+    .arrow-left {
+      left: 0;
+      border-top-right-radius: 50px;
+      border-bottom-right-radius: 50px;
+      z-index: 1;
+    }
+
+    .arrow-right {
+      right: 0;
+      border-top-left-radius: 50px;
+      border-bottom-left-radius: 50px;
+    }
+
+  }
+}
+
+.home-header {
+  min-height: 10vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-top: 5%;
+  margin-bottom: 5%;
+}
+
+.title {
+  font-size: 5rem;
+}
+
+.header-text {
+  color: gray;
+  font-size: 2em;
+}
+
+.header-subtext {
+  color: gray;
+  font-size: 1.5em;
+}
