+++ src/styles/Navbar.scss
@@ -15,9 +15,10 @@ $gray-color: #474747;
     align-items: center;
     padding: 5vh 0;
     font-family: 'Montserrat', sans-serif;
-    font-size: 2rem;
+    font-size: 1.8rem;
     font-weight: 900;
     text-transform: uppercase;
+    z-index: 3;
 
     .navLogo {
         width: 100px;
@@ -28,7 +29,7 @@ $gray-color: #474747;
         display: flex;
         flex-direction: column;
         align-items: flex-end;
-        margin-top: 10vh;
+        margin-top: 8vh;
 
         li {
             list-style: none;
